# 3d-model-server/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "3D模型管理系统 API",
    "description": "3D模型管理系统的API文档，包含用户管理和3D模型管理功能",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/models/{userId}": {
      "get": {
        "summary": "获取用户的所有3D模型",
        "description": "获取指定用户的所有3D模型路径",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["/path/to/model1", "/path/to/model2"]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "无效的用户ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid user ID"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建新的3D模型",
        "description": "为指定用户创建新的3D模型记录",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["model_path"],
                "properties": {
                  "model_path": {
                    "type": "string",
                    "description": "模型文件路径"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "model_path": {
                          "type": "string",
                          "example": "/path/to/model"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "创建失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "用户ID不存在，请检查用户ID是否正确"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
} 